Contract changes for single token zap (#260)
* Contract changes for single token zap

* Apply formatting changes

+++ contracts/interfaces/IWETH.sol
@@ -25,8 +25,8 @@ pragma solidity 0.7.6;
 import "@openzeppelin/contracts/token/ERC20/IERC20.sol";
 
 interface IWETH is IERC20 {
-    event  Deposit(address indexed dst, uint wad);
-    event  Withdrawal(address indexed src, uint wad);
+    event Deposit(address indexed dst, uint256 wad);
+    event Withdrawal(address indexed src, uint256 wad);
 
     function deposit() external payable;
